requirements.txt
opencv-python==4.11.0.86
ffmpeg-python>=0.2.0
python-dotenv==1.0.1
transformers>=4.39.0
huggingface_hub>=0.29.1
timm>=1.0.15
einops>=0.8.1
accelerate>=0.26.0
aiohttp==3.10.11
python_socketio>=5.12.0
torch~=2.6.0+cu124
numpy~=1.26.3
pillow~=11.1.0
loguru~=0.7.3
soundfile~=0.13.1
click~=8.1.8
tqdm~=4.67.1
matplotlib~=3.10.0
torchaudio~=2.6.0+cu124
cachetools~=5.5.2
httpx~=0.28.1
requests~=2.32.3
uvicorn~=0.34.0
pydantic~=2.10.6
rich~=13.9.4
tiktoken~=0.9.0
yt_dlp>=2025.3.21
natsort>=8.4.0
hydra-core>=1.3.2
pytorch-lightning~=2.5.1
lightning~=2.5.1
loralib>=0.1.2
pyrootutils>=1.0.4
vector-quantize-pytorch>=1.14.24
einx[torch]>=0.2.2